requirements.txt
# ChicImages Django Project Requirements
Django==4.2.7
Pillow==10.0.0
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
psycopg2-binary==2.9.9